% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_contingency.R
\name{kable_contingency}
\alias{kable_contingency}
\title{Produces a contingency table from all provided columns
and outputs a kableExtra object for the table}
\usage{
kable_contingency(x)
}
\description{
Produces a contingency table from all provided columns
and outputs a kableExtra object for the table
}
