% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_names_AH.R
\name{make_names_AH}
\alias{make_names_AH}
\title{Make names for df}
\usage{
make_names_AH(chnames)
}
\arguments{
\item{chnames}{vector of strings-to-be-namse}
}
\description{
Similar function to base make.names, but starts from .1
}
