% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_gene_list.R
\name{pathway_gene_list}
\alias{pathway_gene_list}
\title{Retrieve pathway gene list}
\usage{
pathway_gene_list(pathway, gene_format = "ENTREZ")
}
\arguments{
\item{pathway}{pathway name or a keyword}

\item{gene_format}{output format: ENTREZ or SYMBOL. default: ENTREZ}
}
\description{
The function retrieves pathway genes from KEGG and returns either as
entrez genes or gene symbols.
}
