% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_GO_gene_list.R
\name{retrieve_GO_gene_list}
\alias{retrieve_GO_gene_list}
\title{Retrieve genes from GO term in human}
\usage{
retrieve_GO_gene_list(GOquery, gene.format = "SYMBOL")
}
\arguments{
\item{GOquery}{GO id or the name of GO}

\item{gene.format}{output format: ENTREZ or SYMBOL or ENSEMBL. default: SYMBOL}
}
\description{
The function retrieves all genes from GO term and returns either as
entrez genes or gene symbols.
}
